+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
-testpaths = ubc ubcsp
-python_files = ubc/*.py ubcsp/*.py
+testpaths = ubcpdk
+python_files = ubcpdk/*.py
 addopts = '--tb=short'
 
 filterwarnings =
