+++ ubcpdk/simulations/circuits/__init__.py
@@ -1,3 +1,6 @@
+"""Circuit simulations
+"""
+
 from .add_grating_coupler import add_grating_coupler
 from .mzi import mzi
 from .mzi_spectrum import mzi_spectrum
